Added a EMET configuration file to the Repo
THis configuration file is one that has been used in production and
could be used by anyone that wanted to use one. In the future EMET will
actually attempt to self configure at instalation and reconfigure during
the grace period. This will cause a little bit of pain for the users but
in the end will give them the most protection possible with the least
amount of pain.