Added and uninstall script and fixed bugs
The uninstall script is located in EMT\Scripts and will require admin
priv to run.

EMET event converter now properly only gets the intended information and
no long exits out because of thw work being done in the sysmon
integration.

Generate-EmetEvents needs more love but i added in a check for the -Live
switch so that the stock event functionallity still works.